requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.9.0+cu128
torchvision==0.24.0+cu128
torchaudio==2.9.0+cu128

tqdm==4.67.3
numpy==2.2.6
pandas==2.3.3
seaborn==0.13.2
matplotlib==3.10.8
setuptools==65.5.0

gym==0.21.0
wandb==0.25.1
mujoco==3.3.6
hydra-core==1.3.2
dm_control==1.0.34
